requirements.txt
aiofiles==22.1.0
anyio==3.6.2
asyncio==3.4.3
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
et-xmlfile==1.1.0
fastapi==0.88.0
h11==0.14.0
idna==3.4
numpy==1.24.0
openai==0.25.0
openpyxl==3.0.10
pandas==1.5.2
pandas-stubs==1.5.2.221213
Pillow==9.3.0
pydantic==1.10.2
python-dateutil==2.8.2
python-dotenv==0.21.0
python-multipart==0.0.5
pytz==2022.7
requests==2.28.1
six==1.16.0
sniffio==1.3.0
starlette==0.22.0
tqdm==4.64.1
types-pytz==2022.7.0.0
typing-extensions==4.4.0
urllib3==1.26.13
uvicorn==0.20.0
